Move random generators to their own file
This makes it more readable when extracting the common
randomBytes function and consolidates use of some dependencies.
