+++ manageiq-operator/pkg/helpers/miq-components/app-secret.go
@@ -1,10 +1,6 @@
 package miqtools
 
 import (
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/base64"
-
 	miqv1alpha1 "github.com/manageiq/manageiq-pods/manageiq-operator/pkg/apis/manageiq/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -27,17 +23,4 @@ func AppSecret(cr *miqv1alpha1.Manageiq) *corev1.Secret {
 		},
 		StringData: secret,
 	}
-}
-
-func generateEncryptionKey() string {
-	buf := make([]byte, 32)
-	_, err := rand.Read(buf)
-	if err != nil {
-		panic(err) // out of randomness, should never happen
-	}
-
-	h := sha256.New()
-	h.Write(buf)
-
-	return base64.URLEncoding.EncodeToString(h.Sum(nil))
 }
